修正 VectorXlApi 结构体 totalBitCnt 类型为 ushort
将 VectorXlApi 结构体中的 totalBitCnt 字段类型由 byte 改为 ushort，并添加注释说明其在 vxlapi.h 中为 unsigned short，避免声明为 1 字节导致 dlc 与 data 字段错位。

+++ CAN/CanPlugin/Adapters/Vector/VectorXlApi.cs
@@ -96,7 +96,8 @@ internal static class VectorXlApi
         public uint crc;
         [MarshalAs(UnmanagedType.ByValArray, SizeConst = 12)]
         public byte[] reserved3;
-        public byte totalBitCnt;
+        // vxlapi.h 中 totalBitCnt 为 unsigned short；若按 1 字节声明会使 dlc 与 data 整体错位。
+        public ushort totalBitCnt;
         public byte dlc;
         [MarshalAs(UnmanagedType.ByValArray, SizeConst = 5)]
         public byte[] reserved4;
